requirements.txt
requests
beautifulsoup4
dnspython
argparse
tqdm
pyopenssl
lolcat
pyfiglet
